resources/code/workspace/puppy_guide/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==3.0.2",
    "python-dotenv==1.2.1",
]
